# openapi_gpt.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Finance Management API",
    "description": "API pre prístup k finančným dátam - transakcie, účty, kategórie, štatistiky",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "https://finance-management-system-production.up.railway.app"
    }
  ],
  "paths": {
...
    "/api/gpt/transactions/summary": {
      "get": {
        "summary": "Zhrnutie transakcií",
        "description": "Celkové výdavky, príjmy, počet transakcií",
        "operationId": "getTransactionsSummary",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 30
            },
            "description": "Počet dní spätne (default: 30)"
          },
          {
            "name": "account_id",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Filtrovať podľa účtu (nepovinné)"
          }
        ],
        "responses": {
          "200": {
            "description": "Zhrnutie",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "period_days": {"type": "integer"},
                    "summary": {
                      "type": "object",
                      "properties": {
                        "total_count": {"type": "integer"},
                        "total_expenses": {"type": "number"},
                        "total_income": {"type": "number"},
                        "avg_expense": {"type": "number"}
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/gpt/transactions/recent": {
      "get": {
        "summary": "Posledné transakcie",
        "description": "Najnovšie transakcie",
        "operationId": "getRecentTransactions",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Počet transakcií (default: 10)"
          }
        ],
        "responses": {
          "200": {
            "description": "Zoznam transakcií"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API_KEY"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}